자주 찾는 질문 선택 삭제까지 완료

+++ src/main/java/com/playwith/play/domain/qna/repository/QnaRepository.java
@@ -3,5 +3,8 @@ package com.playwith.play.domain.qna.repository;
 import com.playwith.play.domain.qna.entity.Qna;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QnaRepository extends JpaRepository<Qna, Integer> {
+import java.util.List;
+
+public interface QnaRepository extends JpaRepository<Qna, Long> {
+
 }
